% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxrobust.R
\docType{package}
\name{coxrobust}
\alias{coxrobust}
\title{coxrobust Overview}
\description{
This package currently has one main function that fits a robustly proportional
hazards regression model
}
\section{Main Functions}{

\itemize{
\item \code{coxr()}
\item \code{gen_data()}
\item \code{plot.coxr()}
\item \code{predict.coxr()}
}
}

